+++ lib/extensions/head.rs
@@ -5,6 +5,7 @@ use crate::{
 };
 
 /// Middleware to add support for the HTTP [HEAD](https://developer.mozilla.org/en-US/docs/web/http/methods/head) method.
+///
 /// It does this by changing the method to GET and adding a special header (`afire::head`).
 /// Once the response is processed by the normal route handler, the middleware will check if the header is present.
 /// If it is, any body data will be discarded and the [Content-Length] header will be added, if it is not already present.
@@ -25,7 +26,7 @@ impl Head {
     /// Set whether to add the Content-Length header to streaming responses.
     /// ## Example
     /// ```
-    /// # use afire::extension::Head;
+    /// # use afire::extensions::Head;
     /// let head = Head::new().streaming(false);
     /// ```
     pub fn streaming(mut self, streaming: bool) -> Self {