+++ lib/extensions/mod.rs
@@ -3,6 +3,8 @@ pub mod head;
 pub mod logger;
 pub mod ratelimit;
 pub mod real_ip;
+pub mod redirect;
 pub mod request_id;
+pub mod route_shorthands;
 pub mod serve_static;
 pub mod trace;